+++ CHANGELOG.md
@@ -5,7 +5,17 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [2.6.0] - 2024-02-06
+- Added option to determine if server requires a reboot based on current and installed kernel.
 - Updated python-lib to v3.0.2
+
+### Added
+- kernel_run: Checks to see if the kernel check can be done and process the output.
+- kernel_check: Compares the current running kernel version to the latest kernel installed version.
+- get_latest_kernel: Return the latest kernel version.
+- get_running_kernel: Return the running kernel version.
+- get_installed_kernels: Return the installed kernel versions on the server.
+- create_template_dict: Set up dictionary with server-level details.
+- mongo_insert: Insert data into MongoDB.
 
 ### Changed
 - Documentation updates.